% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmask.R
\name{fmask}
\alias{fmask}
\alias{fmask,Raster,sf-method}
\alias{fmask,Raster,Spatial-method}
\alias{fmask,RasterLayer,RasterLayer-method}
\alias{fmask,RasterStackBrick,RasterLayer-method}
\alias{fmask,RasterLayer,RasterStackBrick-method}
\alias{fmask,RasterStackBrick,RasterStackBrick-method}
\title{Fast masking of values in a Raster object}
\usage{
fmask(x, mask, ...)

\S4method{fmask}{Raster,sf}(
  x,
  mask,
  filename = "",
  inverse = FALSE,
  updatevalue = NA,
  updateNA = FALSE,
  ...
)

\S4method{fmask}{Raster,Spatial}(
  x,
  mask,
  filename = "",
  inverse = FALSE,
  updatevalue = NA,
  updateNA = FALSE,
  ...
)

\S4method{fmask}{RasterLayer,RasterLayer}(
  x,
  mask,
  filename = "",
  inverse = FALSE,
  maskvalue = NA,
  updatevalue = NA,
  updateNA = FALSE,
  ...
)

\S4method{fmask}{RasterStackBrick,RasterLayer}(
  x,
  mask,
  filename = "",
  inverse = FALSE,
  maskvalue = NA,
  updatevalue = NA,
  updateNA = FALSE,
  ...
)

\S4method{fmask}{RasterLayer,RasterStackBrick}(
  x,
  mask,
  filename = "",
  inverse = FALSE,
  maskvalue = NA,
  updatevalue = NA,
  updateNA = FALSE,
  ...
)

\S4method{fmask}{RasterStackBrick,RasterStackBrick}(
  x,
  mask,
  filename = "",
  inverse = FALSE,
  maskvalue = NA,
  updatevalue = NA,
  updateNA = FALSE,
  ...
)
}
\arguments{
\item{x}{Raster* object}

\item{mask}{Raster* object or a Spatial* object from sf or sp classes}

\item{...}{additional arguments as in writeRaster}

\item{filename}{character. Optional output filename}

\item{inverse}{logical. If TRUE, areas on mask that are _not_ the maskvalue are masked}

\item{updatevalue}{numeric. The value that cells of x should become if they are 
not covered by mask (and not NA)}

\item{updateNA}{logical. If TRUE, NA values outside the masked area are also 
updated to the the updatevalue (only relevant if the updatevalue is not NA}

\item{maskvalue}{numeric. The value in mask that indicates the cells of x that 
should become updatevalue (default = NA)}
}
\value{
A raster of the same size, extent, resolution and projection as the
provided raster template.
}
\description{
Create a new Raster* object that has the same values as x, except for the cells
that are NA (or other maskvalue) in a 'mask'. These cells become NA (or other
updatevalue). The mask can be either another Raster* object of the same extent
and resolution, or a Spatial* object (e.g. SpatialPolygons) in which case all
cells that are not covered by the Spatial object are set to updatevalue. You
can use inverse=TRUE to set the cells that are not NA (or other maskvalue) in
the mask, or not covered by the Spatial* object, to NA (or other updatvalue).
}
\examples{

}
\references{
Wylie, C., Romney, G., Evans, D., & Erdahl, A. (1967).
  Half-tone perspective drawings by computer. Proceedings of the November
  14-16, 1967, Fall Joint Computer Conference. AFIPS '67 (Fall).
  <https://dx.doi.org/10.1145/1465611.1465619>
}
