% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{fraster-package}
\alias{fraster-package}
\title{Tools for faster raster manipulation}
\description{
\code{fraster} provides tools for performing faster raster manipulations and
calculations.
}
\details{

}
\seealso{
\code{\link[raster]{raster-package}}, \code{\link[sf]{sf}}, \code{\link[sp]{sp}}
}
\author{
D. Scott Rinnan
}
