\name{matrix_add}
\alias{matrix_add}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Matrix Addition Test
}
\description{
  adds two conformable matrices.
}
\usage{
  matrix_add(a,b)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{a}{ Matrix }
  \item{b}{ Matrix }
}
\details{
  TBD.
}
\value{
  Numeric matrix. 
  %%  ~Describe the value returned
  %%  If it is a LIST, use
  %%  \item{comp1 }{Description of 'comp1'}
  %%  \item{comp2 }{Description of 'comp2'}
  %% ...
}
\references{
}

\author{
  Greg Lee
}
\note{
}
%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
  %% \code{\link{matrix_add}}
}
\examples{
\dontrun{
matrix_add(a,b)
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
